Replaced deprecated setup.py with pyproject.toml

+++ mojadata/build_wheel.bat
@@ -1,6 +1,7 @@
 @echo off
-rd /s /q dist
-rd /s /q build
-rd /s /q mojadata.egg-info
-python -m pip install --upgrade setuptools wheel
-python setup.py sdist bdist_wheel
+set PYTHONHOME=c:\python310
+
+if exist dist rd /s /q dist
+python -m pip wheel -w dist --no-deps .
+if exist build rd /s /q build
+if exist mojadata.egg-info rd /s /q mojadata.egg-info
